requirements.txt
absl-py==1.0.0
amqp==5.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==0.13.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
awscli==1.24.7
backcall==0.2.0
beautifulsoup4==4.10.0
bidict==0.22.0
billiard==3.6.4.0
bleach==5.0.0
blis==0.7.7
boto3==1.23.7
botocore==1.26.7
builtwith==1.3.4
cachetools==5.0.0
catalogue==2.0.7
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
chi==0.2
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.0.0
colab==1.13.5
colorama==0.4.4
comtypes==1.1.11
contextlib2==21.6.0
cssselect==1.1.0
cv2-plt-imshow==0.0.1
cycler==0.11.0
cymem==2.0.6
Cython==0.29.30
darr==0.5.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diazo==1.4.2
distlib==0.3.4
Django==1.7.11
django-haystack==2.4.1
django-hitcounter==0.1.1
django-revproxy==0.10.0
django-taggit==0.22.2
docker==4.2.2
docutils==0.16
dominate==2.6.0
english-words==1.1.0
entrypoints==0.4
et-xmlfile==1.1.0
etiquetando==0.1
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
Flask==2.1.0
Flask-Bootstrap==3.3.7.1
Flask-SocketIO==5.1.1
flatbuffers==2.0
fonttools==4.31.1
future==0.18.2
gast==0.5.3
gensim==4.1.2
Geometry==0.0.23
google==3.0.0
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
gym==0.23.1
gym-notices==0.0.6
h5py==3.6.0
html2text==2020.1.16
idna==3.3
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.1.0
Js2Py==0.71
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kaggle==1.5.12
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==8.7
keyrings.alt==4.1.0
kiwisolver==1.4.0
kombu==5.2.4
labelImg==1.8.6
langcodes==3.3.0
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mediapipe==0.8.9.1
mistune==0.8.4
murmurhash==1.0.6
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.4
nltk==3.7
notebook==6.4.11
numpy==1.22.3
oauthlib==3.2.0
object-detection==0.0.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opendatasets==0.1.20
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
patsy==0.5.2
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.0.1
Plate==1.0.0
platformdirs==2.5.1
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
pure-eval==0.2.2
py2exe==0.11.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycodestyle==2.8.0
PyContracts==1.8.12
pycparser==2.21
pydantic==1.8.2
pygame==2.1.2
PyGeometry==1.5.6
Pygments==2.11.2
pyjsparser==2.7.1
PyMySQL==1.0.2
pyparsing==3.0.7
pypiwin32==223
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-engineio==4.3.1
python-slugify==6.1.1
python-socketio==5.5.2
python-whois==0.7.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
redis==4.2.0
regex==2022.3.15
repoze.lru==0.7
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
Sastrawi==1.0.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sound==0.1.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
split-folders==0.5.1
SQLAlchemy==1.4.36
srsly==2.4.2
stack-data==0.2.0
statsmodels==0.13.2
stemming==1.0.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.15.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.15
threadpoolctl==3.1.0
time-convert==1.2.5
tinycss2==1.1.1
toml==0.10.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typer==0.4.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
vine==5.0.0
virtualenv==20.14.1
visitor==0.1.3
voluptuous==0.13.1
wasabi==0.9.0
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.0.3
wget==3.2
Whoosh==2.7.4
widgetsnbextension==3.6.0
wordcloud==1.8.1
wrapt==1.14.0
yolo==0.3.1
